+++ src/svelte/components/checkbox.svelte
@@ -29,6 +29,7 @@
 
   function onChange(event) {
     emit('change', [event]);
+    checked = event.target.checked;
   }
 
   onMount(() => {