+++ src/svelte/components/list-input.svelte
@@ -245,6 +245,9 @@
     if (!(validateOnBlur || validateOnBlur === '') && (validate || validate === '') && inputEl) {
       validateInput(inputEl);
     }
+    if (inputEl && type !== 'texteditor' && type !== 'colorpicker' && type !== 'datepicker') {
+      value = inputEl.value;
+    }
   }
 
   function onFocus(...args) {
@@ -267,6 +270,7 @@
 
     if (type === 'texteditor') {
       emit('textEditorChange', [args[1]]);
+      value = args[1];
     }
   }
 
@@ -291,6 +295,7 @@
           on: {
             change(calendar, calendarValue) {
               emit('calendarChange', [calendarValue]);
+              value = calendarValue;
             },
           },
           ...(calendarParams || {}),
@@ -303,6 +308,7 @@
           on: {
             change(colorPicker, colorPickerValue) {
               emit('colorPickerChange', [colorPickerValue]);
+              value = colorPickerValue;
             },
           },
           ...(colorPickerParams || {}),