+++ src/svelte/components/list-item.svelte
@@ -243,6 +243,7 @@
   }
   function onChange(event) {
     emit('change', [event]);
+    checked = event.target.checked;
   }
 
   useSmartSelect(