+++ src/svelte/components/radio.svelte
@@ -28,6 +28,7 @@
 
   function onChange(event) {
     emit('change', [event]);
+    checked = event.target.checked;
   }
 </script>