+++ src/vue/components/radio.vue
@@ -32,11 +32,12 @@ export default {
     readonly: Boolean,
     ...colorProps,
   },
-  emits: ['update:checked'],
+  emits: ['update:checked', 'change'],
   setup(props, { emit }) {
     const inputElRef = ref(null);
 
     const onChange = (event) => {
+      emit('change', event);
       emit('update:checked', event.target.checked);
     };